Generating origin and destination vectors for orthographic camera and rotation (#3)
* Renamed: Camera.hp -> CameraInfo.hpp
Removed: OrthographicCamera.hpp struct
Added: orthographicCamera.cpp function
Added: Vector2 header and implementation
Added: Size header and implementation
Refactored: ppm.cpp
Added: more filters
Added: RotateableObject header

* Simplified ppm export functions

* Simplified ppm export functions

* Added: Vector3 addition and substraction with double
Added: matrix rotation for 3D
Added: orthographic camera

* Added: preludium to domain config
Added: default constructor for Size
Added: header for orthographicCamera
Changed: destination to EPSILON*100
Changed: bmp_headers array size to macro defined value
Changed: BMP_ALL_HEADERS_BYTES macro to be computed
Changed: added public inheritance for Object and related

+++ Simple CPP Ray Tracer/src/ppm.cpp
@@ -22,7 +22,7 @@ void exportPPMtxt(RGB *data, Size size)
 	fprintf(file, "P%d\n%d %d\n%d\n", 3, size.width, size.height, 255);
 	for (uint y = 0; y < size.height; y++)
 		for (uint x = 0; x < size.width; x++) {
-			RGB t = data[y * (int)size.height + x];
+			RGB t = data[y * size.height + x];
 			fprintf(file, "%d %d %d\n", t.R, t.G, t.B);
 		}
 }
@@ -42,7 +42,7 @@ void exportPPMbin(RGB* data, Size size)
 	fprintf(file, "P%d\r%d %d\r%d\r", 6, size.width, size.height, 255);
 	for (uint y = 0; y < size.height; y++)
 		for (uint x = 0; x <= size.width; x++) {
-			RGB t = data[y * (int)size.height + x];
+			RGB t = data[y * size.height + x];
 			fprintf(file, "%c%c%c", t.R, t.G, t.B);
 		}
 }
